+++ client/CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Changed the date facets to full dates for facets and advanced search ([#39](https://github.com/project-lux/lux-frontend/issues/39)).
+
 ### Removed
 
 ### Fixed